--- lib/question.rb
@@ -1,16 +1,0 @@
-module Question
-
-  def question(question, a1, a2)
-    puts question
-    puts a1.upcase + '/' + a2.upcase
-    answer = gets.chomp.downcase
-    if answer == a1.downcase
-      true
-    elsif answer == a2.downcase
-      return false
-    else
-      question(question, a1, a2)
-    end
-  end
-
-end
